Specify dtype for numpy arrays

+++ tests/test_3209_awkwardforth_read_negative_number_of_items.py
@@ -9,11 +9,11 @@ import awkward as ak
 
 def test_read_negative_number_of_items():
     vm = ak.forth.ForthMachine32("input source -5 source #q-> stack")
-    vm.run({"source": np.array([1, 2, 3, 4, 5])})
+    vm.run({"source": np.array([1, 2, 3, 4, 5], dtype=np.int64)})
     assert vm.stack == []
 
     vm = ak.forth.ForthMachine32("input source output sink float64 -5 source #q-> sink")
-    vm.run({"source": np.array([1, 2, 3, 4, 5])})
+    vm.run({"source": np.array([1, 2, 3, 4, 5], dtype=np.int64)})
     assert vm.output("sink").tolist() == []
 
 
@@ -21,13 +21,13 @@ def test_read_negative_and_positive_number_of_items():
     vm = ak.forth.ForthMachine32(
         "input source -5 source #q-> stack 5 source #q-> stack"
     )
-    vm.run({"source": np.array([1, 2, 3, 4, 5])})
+    vm.run({"source": np.array([1, 2, 3, 4, 5], dtype=np.int64)})
     assert vm.stack == [1, 2, 3, 4, 5]
 
     vm = ak.forth.ForthMachine32(
         "input source output sink float64 -5 source #q-> sink 5 source #q-> sink"
     )
-    vm.run({"source": np.array([1, 2, 3, 4, 5])})
+    vm.run({"source": np.array([1, 2, 3, 4, 5], dtype=np.int64)})
     assert vm.output("sink").tolist() == [1, 2, 3, 4, 5]
 
 
@@ -35,11 +35,11 @@ def test_read_positive_and_negative_number_of_items():
     vm = ak.forth.ForthMachine32(
         "input source 5 source #q-> stack -5 source #q-> stack"
     )
-    vm.run({"source": np.array([1, 2, 3, 4, 5])})
+    vm.run({"source": np.array([1, 2, 3, 4, 5], dtype=np.int64)})
     assert vm.stack == [1, 2, 3, 4, 5]
 
     vm = ak.forth.ForthMachine32(
         "input source output sink float64 5 source #q-> sink -5 source #q-> sink"
     )
-    vm.run({"source": np.array([1, 2, 3, 4, 5])})
+    vm.run({"source": np.array([1, 2, 3, 4, 5], dtype=np.int64)})
     assert vm.output("sink").tolist() == [1, 2, 3, 4, 5]
